# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CyberLab Portal API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Lab {
  id: number;
  name: string;
  description: string;
  category: string;
  difficulty: string;
  pricePerHour: number;
  maxHours: number;
  tools: string[];
  /** @nullable */
  imageUrl: string | null;
  isActive: boolean;
  createdAt: string;
}

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  pending: "pending",
  paid: "paid",
  provisioned: "provisioned",
  expired: "expired",
  failed: "failed",
} as const;

export interface Booking {
  id: number;
  userId: string;
  labId: number;
  hours: number;
  totalAmount: number;
  status: BookingStatus;
  /** @nullable */
  paymentTxnId: string | null;
  /** @nullable */
  labAccessUrl: string | null;
  /** @nullable */
  labCredentials: string | null;
  /** @nullable */
  provisionedAt: string | null;
  /** @nullable */
  expiresAt: string | null;
  createdAt: string;
  lab: Lab;
}

export interface CreateBookingBody {
  labId: number;
  /**
   * @minimum 1
   * @maximum 24
   */
  hours: number;
}

export interface ProvisionResult {
  success: boolean;
  accessUrl: string;
  credentials: string;
  expiresAt: string;
}

export interface InitiatePaymentBody {
  bookingId: number;
}

export interface PaymentInitiateResponse {
  accessKey: string;
  paymentUrl: string;
  txnId: string;
  amount: number;
  productInfo: string;
  name: string;
  email: string;
}

export interface PaymentCallbackResponse {
  status: string;
  /** @nullable */
  bookingId: number | null;
}

export interface DashboardSummary {
  totalBookings: number;
  activeBookings: number;
  totalSpent: number;
  completedLabs: number;
  recentBookings: Booking[];
}

export type PaymentCallbackBody = { [key: string]: unknown };