# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CyberLab Portal API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all available labs
 */
export const GetLabsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string(),
  category: zod.string(),
  difficulty: zod.string(),
  pricePerHour: zod.number(),
  maxHours: zod.number(),
  tools: zod.array(zod.string()),
  imageUrl: zod.string().nullable(),
  isActive: zod.boolean(),
  createdAt: zod.coerce.date(),
});
export const GetLabsResponse = zod.array(GetLabsResponseItem);

/**
 * @summary Get a single lab
 */
export const GetLabParams = zod.object({
  id: zod.coerce.number(),
});

export const GetLabResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string(),
  category: zod.string(),
  difficulty: zod.string(),
  pricePerHour: zod.number(),
  maxHours: zod.number(),
  tools: zod.array(zod.string()),
  imageUrl: zod.string().nullable(),
  isActive: zod.boolean(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Get all bookings for the current user
 */
export const GetMyBookingsResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  labId: zod.number(),
  hours: zod.number(),
  totalAmount: zod.number(),
  status: zod.enum(["pending", "paid", "provisioned", "expired", "failed"]),
  paymentTxnId: zod.string().nullable(),
  labAccessUrl: zod.string().nullable(),
  labCredentials: zod.string().nullable(),
  provisionedAt: zod.coerce.date().nullable(),
  expiresAt: zod.coerce.date().nullable(),
  createdAt: zod.coerce.date(),
  lab: zod.object({
    id: zod.number(),
    name: zod.string(),
    description: zod.string(),
    category: zod.string(),
    difficulty: zod.string(),
    pricePerHour: zod.number(),
    maxHours: zod.number(),
    tools: zod.array(zod.string()),
    imageUrl: zod.string().nullable(),
    isActive: zod.boolean(),
    createdAt: zod.coerce.date(),
  }),
});
export const GetMyBookingsResponse = zod.array(GetMyBookingsResponseItem);

/**
 * @summary Create a new booking
 */
export const createBookingBodyHoursMax = 24;

export const CreateBookingBody = zod.object({
  labId: zod.number(),
  hours: zod.number().min(1).max(createBookingBodyHoursMax),
});

/**
 * @summary Get a single booking
 */
export const GetBookingParams = zod.object({
  id: zod.coerce.number(),
});

export const GetBookingResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  labId: zod.number(),
  hours: zod.number(),
  totalAmount: zod.number(),
  status: zod.enum(["pending", "paid", "provisioned", "expired", "failed"]),
  paymentTxnId: zod.string().nullable(),
  labAccessUrl: zod.string().nullable(),
  labCredentials: zod.string().nullable(),
  provisionedAt: zod.coerce.date().nullable(),
  expiresAt: zod.coerce.date().nullable(),
  createdAt: zod.coerce.date(),
  lab: zod.object({
    id: zod.number(),
    name: zod.string(),
    description: zod.string(),
    category: zod.string(),
    difficulty: zod.string(),
    pricePerHour: zod.number(),
    maxHours: zod.number(),
    tools: zod.array(zod.string()),
    imageUrl: zod.string().nullable(),
    isActive: zod.boolean(),
    createdAt: zod.coerce.date(),
  }),
});

/**
 * @summary Provision the lab for a paid booking
 */
export const ProvisionLabParams = zod.object({
  id: zod.coerce.number(),
});

export const ProvisionLabResponse = zod.object({
  success: zod.boolean(),
  accessUrl: zod.string(),
  credentials: zod.string(),
  expiresAt: zod.coerce.date(),
});

/**
 * @summary Initiate Easebuzz payment for a booking
 */
export const InitiatePaymentBody = zod.object({
  bookingId: zod.number(),
});

export const InitiatePaymentResponse = zod.object({
  accessKey: zod.string(),
  paymentUrl: zod.string(),
  txnId: zod.string(),
  amount: zod.number(),
  productInfo: zod.string(),
  name: zod.string(),
  email: zod.string(),
});

/**
 * @summary Easebuzz payment callback
 */
export const PaymentCallbackBody = zod.record(zod.string(), zod.unknown());

export const PaymentCallbackResponse = zod.object({
  status: zod.string(),
  bookingId: zod.number().nullable(),
});

/**
 * @summary Get dashboard stats for the current user
 */
export const GetDashboardSummaryResponse = zod.object({
  totalBookings: zod.number(),
  activeBookings: zod.number(),
  totalSpent: zod.number(),
  completedLabs: zod.number(),
  recentBookings: zod.array(
    zod.object({
      id: zod.number(),
      userId: zod.string(),
      labId: zod.number(),
      hours: zod.number(),
      totalAmount: zod.number(),
      status: zod.enum(["pending", "paid", "provisioned", "expired", "failed"]),
      paymentTxnId: zod.string().nullable(),
      labAccessUrl: zod.string().nullable(),
      labCredentials: zod.string().nullable(),
      provisionedAt: zod.coerce.date().nullable(),
      expiresAt: zod.coerce.date().nullable(),
      createdAt: zod.coerce.date(),
      lab: zod.object({
        id: zod.number(),
        name: zod.string(),
        description: zod.string(),
        category: zod.string(),
        difficulty: zod.string(),
        pricePerHour: zod.number(),
        maxHours: zod.number(),
        tools: zod.array(zod.string()),
        imageUrl: zod.string().nullable(),
        isActive: zod.boolean(),
        createdAt: zod.coerce.date(),
      }),
    }),
  ),
});