# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CyberLab Portal API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from "./bookingStatus";
import type { Lab } from "./lab";

export interface Booking {
  id: number;
  userId: string;
  labId: number;
  hours: number;
  totalAmount: number;
  status: BookingStatus;
  /** @nullable */
  paymentTxnId: string | null;
  /** @nullable */
  labAccessUrl: string | null;
  /** @nullable */
  labCredentials: string | null;
  /** @nullable */
  provisionedAt: Date | null;
  /** @nullable */
  expiresAt: Date | null;
  createdAt: Date;
  lab: Lab;
}