# lib/api-zod/src/generated/types/createBookingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CyberLab Portal API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBookingBody {
  labId: number;
  /**
   * @minimum 1
   * @maximum 24
   */
  hours: number;
}