# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CyberLab Portal API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Booking } from "./booking";

export interface DashboardSummary {
  totalBookings: number;
  activeBookings: number;
  totalSpent: number;
  completedLabs: number;
  recentBookings: Booking[];
}