# lib/api-zod/src/generated/types/initiatePaymentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CyberLab Portal API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InitiatePaymentBody {
  bookingId: number;
}