# lib/api-zod/src/generated/types/lab.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CyberLab Portal API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Lab {
  id: number;
  name: string;
  description: string;
  category: string;
  difficulty: string;
  pricePerHour: number;
  maxHours: number;
  tools: string[];
  /** @nullable */
  imageUrl: string | null;
  isActive: boolean;
  createdAt: Date;
}