# lib/api-zod/src/generated/types/paymentCallbackResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CyberLab Portal API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentCallbackResponse {
  status: string;
  /** @nullable */
  bookingId: number | null;
}