# lib/api-zod/src/generated/types/paymentInitiateResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CyberLab Portal API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentInitiateResponse {
  accessKey: string;
  paymentUrl: string;
  txnId: string;
  amount: number;
  productInfo: string;
  name: string;
  email: string;
}